duy last commit

+++ src/main/java/com/bikiegang/ridesharing/pojo/response/PlannedTripShortDetailResponse.java
@@ -9,7 +9,7 @@ public class PlannedTripShortDetailResponse {
     private String startAddress;
     private String endAddress;
     private double unitPrice;
-    private long ownerDistance;
+    private double ownerDistance;
     private int role;
 
     public PlannedTripShortDetailResponse() {
@@ -64,11 +64,11 @@ public class PlannedTripShortDetailResponse {
         this.unitPrice = unitPrice;
     }
 
-    public long getOwnerDistance() {
+    public double getOwnerDistance() {
         return ownerDistance;
     }
 
-    public void setOwnerDistance(long ownerDistance) {
+    public void setOwnerDistance(double ownerDistance) {
         this.ownerDistance = ownerDistance;
     }
